Revert "Merge remote-tracking branch 'refs/remotes/upperstream/master'"
This reverts commit 10059bd73e4107117e9acfb1bc457ed215b6a837, reversing
changes made to ca3ae4dbd4e4ad76878f5f327ce4a65011a19be2.

+++ examples/core/core_input_mouse_wheel.c
@@ -36,7 +36,7 @@ int main(void)
     {
         // Update
         //----------------------------------------------------------------------------------
-        boxPositionY -= (int)(GetMouseWheelMove()*scrollSpeed);
+        boxPositionY -= (GetMouseWheelMove()*scrollSpeed);
         //----------------------------------------------------------------------------------
 
         // Draw
